requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.8
distro==1.9.0
duckduckgo_search==7.3.0
gpt4all==2.8.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
openai==1.63.2
primp==0.11.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
